soap_architecture/python/backend/portal_service/requirements.txt
asgiref==3.7.2
greenlet==3.1.0
lxml==5.3.0
mysql-connector-python==9.0.0
mysqlclient==2.2.0
packaging==23.1
pyparsing==3.0.9
python-dotenv==1.0.0
pytz==2024.2
spyne==2.14.0
SQLAlchemy==2.0.35
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2023.3